+++ aeronet/objects/include/aeronet/tchars.hpp
@@ -9,7 +9,7 @@ namespace aeronet {
 ///                  / DIGIT / ALPHA
 constexpr bool is_tchar(unsigned char uc) noexcept {
   // Two 64-bit chunks: [0–63], [64–127]
-  static constexpr uint64_t bitmap[2] = {
+  static constexpr uint64_t bitmap[]{
       // low 64
       (1ULL << '!') | (1ULL << '#') | (1ULL << '$') | (1ULL << '%') | (1ULL << '&') | (1ULL << '\'') | (1ULL << '*') |
           (1ULL << '+') | (1ULL << '-') | (1ULL << '.') | (0x3FFULL << '0'),  // digits 0–9
@@ -18,9 +18,10 @@ constexpr bool is_tchar(unsigned char uc) noexcept {
       (0x3FFFFFFULL << ('A' - 64)) |                     // A–Z
           (1ULL << ('^' - 64)) | (1ULL << ('_' - 64)) |  // ^ and _
           (0x3FFFFFFULL << ('a' - 64)) |                 // a–z
-          (1ULL << ('`' - 64)) | (1ULL << ('|' - 64)) | (1ULL << ('~' - 64))};
+          (1ULL << ('`' - 64)) | (1ULL << ('|' - 64)) | (1ULL << ('~' - 64)),
+  };
 
-  return uc < 128U && ((bitmap[uc >> 6] >> (uc & 63)) & 1U) != 0U;
+  return uc < 128U && ((bitmap[uc >> 6U] >> (uc & 63U)) & 1U) != 0U;
 }
 
 constexpr bool is_tchar(char ch) noexcept { return is_tchar(static_cast<unsigned char>(ch)); }